move __main__.py to commands/__main__.py
