+++ pyinstaller/main.py
@@ -1,4 +1,4 @@
-from mapillary_tools.__main__ import main
+from mapillary_tools.commands.__main__ import main
 
 if __name__ == '__main__':
     main()
